% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualpalr.R
\name{show_qualpal}
\alias{show_qualpal}
\title{Display a colour palette}
\usage{
show_qualpal(x, labels = NULL, border = "white", cex = 0.8, ...)
}
\arguments{
\item{x}{A character vector of colours.}

\item{labels}{Optional labels to display for each colour. If \code{NULL} and \code{x}
is named, the names of \code{x} are used. If \code{NULL} and \code{x} is unnamed, no
labels are shown.}

\item{border}{Colour used for rectangle borders.}

\item{cex}{Axis label size.}

\item{...}{Additional arguments passed to \code{\link[graphics:title]{graphics::title()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Displays a palette as a series of coloured rectangles.

This function can be used with both unnamed colour vectors, such as those
returned by \code{\link[=pal_qualpal]{pal_qualpal()}}, and named colour vectors, such as those returned
by \code{\link[=map_qualpal]{map_qualpal()}}.
}
\details{
If \code{x} is a named vector and \code{labels = NULL}, the names are used as labels.
This is particularly useful for palettes generated with \code{\link[=map_qualpal]{map_qualpal()}}.
}
\examples{
# Unnamed palette
pal <- pal_qualpal()(6)
show_qualpal(pal)

# Named palette
x <- c("A", "B", "C", "A")
pal <- map_qualpal(x)
show_qualpal(pal)

}
